Update tests to the correct expected behavior.

+++ src/EditorFeatures/CSharpTest/CodeActions/ExtractMethod/ExtractMethodTests.cs
@@ -594,7 +594,7 @@ class Program
     static void Main()
     {
         byte z = 0;
-        Goo({|Rename:NewMethod|}(), y => (byte)0, z, z);
+        Goo({|Rename:NewMethod|}(), y => 0, z, z);
     }
 
     private static Func<byte, byte> NewMethod()
@@ -633,7 +633,7 @@ class Program
     static void Main()
     {
         byte z = 0;
-        Goo({|Rename:NewMethod|}(), y => { return (byte)0; }, z, z);
+        Goo({|Rename:NewMethod|}(), y => { return 0; }, z, z);
     }
 
     private static Func<byte, byte> NewMethod()
